requirements.txt
Flask
mcstatus
requests
discord.py